lecture10

+++ lecture3/ex1.py
@@ -1,6 +1,6 @@
-input1= float(input("Enter first socre: "))
-input2 = float(input("Enter second socre: "))
-input3 = float(input("Enter third socre: "))
+input1 = int(input("Enter first socre: "))
+input2 = int(input("Enter second socre: "))
+input3 = int(input("Enter third socre: "))
 avg = (input1 + input2 + input3) / 3
 print("The average is:", avg)
 if avg >= 95:
